# demo-usecase/api/collections.openapi.yaml
openapi: 3.1.0
info:
  title: mopay collections-api
  version: "1.0.0"
  description: |
    Collections intake, channel callbacks and the read APIs behind the console
    (PRD §5.1, §5.2), at the scope agreed in `docs/SCOPE.md`.

    This component also owns the continuously-updated **settlement lag** metric
    (README hard constraint 4, OBS-4). The lag signal must keep being emitted when
    `settlement-worker` is not running — a signal that dies with the worker cannot tell
    you the worker died — so it is computed here on a short interval, not by the worker.

    ## Authentication

    Dev-mode bearer tokens, not OIDC. OIDC, identity providers, login screens and session
    management are README.md non-goals (`docs/SCOPE.md` conflict 5).

        Authorization: Bearer dev.<base64url({"sub":"...","merchantId":"...","role":"..."})>

    **Authorization is not dev-mode.** C-6.2 to C-6.7 are enforced server-side on every
    request: merchant scoping comes from the token, never from a query parameter, and the
    console hiding a screen is explicitly not a control (C-6.6).

    Money is integer minor units with an explicit currency. Timestamps are UTC.
servers:
  - url: http://localhost:8088
    description: Local Compose
...
paths:
...
  /v1/merchants/me:
    get:
      tags: [read]
      summary: The authenticated merchant, with the residency statement (S-5, S-5.1)
      operationId: getMerchant
      description: |
        `dataRegionStatement` is plain language derived from the merchant's stored
        `dataRegion`, so S-5.1's claim is true by construction rather than by policy
        (RES-7). The payout account number is masked (OBS-7).
      responses:
        "200":
          description: The merchant
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant: { $ref: "#/components/schemas/Merchant" }
                  dataRegion: { type: string, enum: [KE, NG] }
                  dataRegionStatement:
                    type: string
                    examples: ["Your transaction data is stored in Kenya."]
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    devBearer:
      type: http
      scheme: bearer
      description: |
        Dev-mode token: `dev.<base64url({"sub","merchantId","role"})>`.
        Roles: `merchant`, `service`, `finance`, `operations`.
  responses:
...
    Unauthorized:
      description: No token, or the token is malformed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - unprocessable
            - internal_error
            - upstream_unavailable
            - float_limit_exceeded
            - ledger_not_balanced
            - append_only_violation
            - invalid_status_transition
            - unsupported_channel
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Merchant:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        country: { type: string, enum: [KE, NG] }
        dataRegion:
          type: string
          enum: [KE, NG]
          description: Immutable after creation (RES-5), enforced by a database trigger.
        payoutBank: { type: string }
        payoutAccountName: { type: string }
        payoutAccountNumberMasked:
          type: string
          description: Masked. The full account number never leaves the datastore (OBS-7).
        floatLimitMinor: { type: integer, format: int64 }
        floatLimitCurrency: { type: string, enum: [KES, NGN] }
...
security:
  - devBearer: []